Added a facility to generate a standalone executable for the model (#156)
Summary:
Sometimes, it would be more convenient to debug with a standalone executable without going through the python world.

This change added an option, gen_standalone, to AITDebugSettings, which can be passed to compile_model method. When gen_standalone is True, a standalone executable will be generated in the same folder as test.so, along with suitable changes to the Makefile.

Note that because we aim for assisting debugging, we make a number of simplifications to the entry standalone.cpp:
  * we use the maximum input shapes;
  * we only generate random inputs with a fixed seed;
  * we assume that outputs exist on the host;
  * we disable graph_mode;
  * etc... Once the file is copied into the intemediate working dir (e.g., ./tmp/gen_standalone_fp16) along with other files, users are free to make any changes to the code and run "make" to re-generate the executable.

Pull Request resolved: https://github.com/facebookincubator/AITemplate/pull/156

Reviewed By: ipiszy

Differential Revision: D42855268

Pulled By: chenyang78

fbshipit-source-id: 08357b8c98e4cb78ef2c0649905b5b1d81112a0b