Foundations for linking the constant folding code into the .so (#1179)
Summary:
X-link: https://github.com/fairinternal/AITemplate/pull/1179

* Implement a modified `constant_folding` pass that links the code into the .so instead of doing it all at compile time.
* Add all codegen & C++ changes to support load-time constant folding and constant updating.

This diff is ready for review. However, it is **not to be landed without additional stacked diffs**. I'm splitting up the changes to avoid huge reviews. But we don't want to have two versions of constant folding around, we should delete the old pass when everything is ready.

Reviewed By: chenyang78

Differential Revision: D41806880

fbshipit-source-id: 7884d274df02b24494af7e49f2d16f29695d40ea