place slice_reshape_scatter's output with a suitable offset (#207)
Summary:
This PR actually fixed two issues:

(1) Previously, there was an issue related to place the generated slice_reshape_scatter into a wrong postion in the output, because we did not take the relevant offset into account. This PR fixed the problem by simply introducing an output_offset argument to the relevant kernel. Please check the comment in the gen_function body of slice_reshape_scatter_common for the rationales.

(2) When we update the newly-created slice_reshape_scatter op, we should add it into the "dst_ops" set of its "inputs" rather than overwriting the entire set.

Pull Request resolved: https://github.com/facebookincubator/AITemplate/pull/207

Reviewed By: tissue3

Differential Revision: D43058204

Pulled By: chenyang78

fbshipit-source-id: 65937290b9614a1ca2d48893cae5fa1453272a20