Amend get_item with list of indices to use ait fusion pass (#229)
Summary:
Pull Request resolved: https://github.com/facebookincubator/AITemplate/pull/229

Current implementation of get_item over a list of indices first perform tensor_slice on pieces of index, squeeze, unsqueeze, amd then concatenate.
This is because tensor_slice will squeeze tensor when encounter int, while it cannot handle list of int.

It doesn't work well with AIT's fusion pass because of extra squeeze and unsqueeze.

To avoid this, we want tensor_slice to skip squeeze when inputs is
```
a=torch.randn([10,8])
a[:, [2]]
```
This diff change tensor_slice to be able to handle list of index when `len(list)==1`. It does the same thing as non index version; the only difference is instead of giving a `int`, now we get a `slice(2,3,None).`
Then AIT fusion pass can handle this dynamic shape issue.

Another thing is, in practice, there are quite some dynamic_slice context that can be merged
e.g. `a[0, 2, 3, 4]`, we want to slice twice, one with `slice(0,1,None)`, the other with `slice(2,5,None)`. This fix the issue of subsequent large slices, e.g. https://fburl.com/phabricator/t26ylhbw

Reviewed By: wushirong

Differential Revision: D43071233

fbshipit-source-id: 828a93c9411ddbf19b0717d60469307ec7629c39