allow users to pick the optimization level for the host code (#239)
Summary:
Pull Request resolved: https://github.com/facebookincubator/AITemplate/pull/239

It is typical in our situation that an option --optimize <level> (-Ox) in a Makefile is for a HOST compiler. And the default -O3 does literally nothing except for the enormous compilation time.

So, it is safe to allow users to override this option in order to **significantly speedup** the compilation / testing, especially for very large models.

Reviewed By: frankgt40

Differential Revision: D43189546

fbshipit-source-id: 6b243e3bee48b38a26b9001b9038715aa6c28eb6