allow users to pick the optimization level for the host code (#239)
Summary:
Pull Request resolved: https://github.com/facebookincubator/AITemplate/pull/239

It is typical in our situation that an option --optimize <level> (-Ox) in a Makefile is for a HOST compiler. And the default -O3 does literally nothing except for the enormous compilation time.

So, it is safe to allow users to override this option in order to **significantly speedup** the compilation / testing, especially for very large models.

Use AIT_COMPILER_OPT environment variable, for example "AIT_COMPILER_OPT=-O1".

Reviewed By: frankgt40, chenyang78

Differential Revision: D43189546

fbshipit-source-id: 4d67cf295bf31b1053fe10816fb5646c4dda1f15