Added an environment variable AIT_FORCE_PROFILER_CACHE (#255)
Summary:
Pull Request resolved: https://github.com/facebookincubator/AITemplate/pull/255

When we update our cache versions or make some changes, we may generate
a different cache key for a problem size from the previous version.
Such cache misses can result in re-compiling profilers and re-performing
profiling.

We introduce a new environment variable AIT_FORCE_PROFILER_CACHE, which
may be used to ensure such cache missed to not be overlooked.

Reviewed By: aakhundov

Differential Revision: D43242228

fbshipit-source-id: 038b221fb9fbcde17f7973c3ea4d6b13a32b0e85