Config the mem_align for the Attention kernel (#301)
Summary:
Pull Request resolved: https://github.com/facebookincubator/AITemplate/pull/301

Let's not force the mem_align to be true for the attention kernel.
Instead, we set it to be true when the head_size meets the alignment
requirement and false otherwise. This change allows us to keep
the mem_align benefit while extending the kernel for more cases,
e.g. we just use the minimal alignment value for "ill-aligned"
head_size.

This change also replaced the simple return with a runtime exception
when the alignment check failed.

The temporary alignment check in fx2ait was removed since we
can ensure the kernel safety now.

Differential Revision: D43415811

fbshipit-source-id: 94e3d3c39869a257c64ab25c6211c28c589a4515