make ProfilerMemoryPool work with low-memory cases (#310)
Summary:
Pull Request resolved: https://github.com/facebookincubator/AITemplate/pull/310

Previously, the minimal number of copies in a ProfilerMemoryPool
instance is 2. It does not work for edge cases where we are
under GPU memory pressure. For example, for some models with large
weights being loaded into the GPU global memory, there may be
a very small amount of memory left for the profiler memory pool.
For example, the free memory could only hold for a single copy
or even less.

This PR mitigate the issue by allocating a single shared blob
that is large enough to hold the largest input of the op
being profiled.

Reviewed By: alexanderguzhva

Differential Revision: D43490681

fbshipit-source-id: c84d05bb7a49da5ae8c167d3a270a5d6a3912a44