make better vectorization for "broadcastable" fused_elementwise ops
Previously, we were quite conservative in terms of determining a vector size
for reading operands in the fused_elementwise kernel. We always based on
the minimal alignment value among all the input operands. This particularly
hurted the kernel performance when the last dimension was "broadcastable".
For example, for two input shapes [batch_size, 10, 1] and [batch_size, 10, 8],
we set read_t (i.e. vector read type) to "half", assuming the input type is
float16. It would be more effecient to have a seperate read_t for each input.
We could load the first input with "half" and the second with "uint4".
Furthermore, we could make a vector that contained eight copies of the first input
and add this vector with the uint4 vector value loaded from the second operand.

This PR implemented such an optimization. Performance results showed that it
may result in upto ~2X performance improvements for some cases.