Add jagged Tensor support to fused_elementwise back-end (#342)
Summary:
Pull Request resolved: https://github.com/facebookincubator/AITemplate/pull/342

In this diff, the `fused_elementwise` back-end is extended to support mixed jagged Tensor / dense Tensor inputs, along with all the pre-existing features of the back-end (broadcasting, vectorized I/O, nested expression generation, etc.). The added functionality should be orthogonal to the existing one. The diff also adds a few Python / PyTorch-based tools in the `testing/jagged_utils.py` that can be used for reference computation in future jagged-aware ops' unit tests.

## Implementation Details

When `fused_elementwise` has are only jagged inputs, only dense inputs, or the highest-rank dense input shape does not overlap with the jagged dimensions of the jagged inputs (i.e., with the `JaggedIntVar` in the jagged input shape), all jagged inputs are treated as dense (with the first `total_length` dimension followed by non-jagged dimensions), and the implementation relies on the pre-existing back-end implementation for simplicity and efficiency.

For the cases when the shape of at least one of the dense inputs overlaps with the jagged dimensions of the jagged inputs, the semantics of the elementwise operation is that the elements of the jagged inputs are used in computation together with the elements of the dense inputs at the corresponding positions. Consequently, the elements in the dense inputs not having the counterparts in the jagged inputs are ignored. (It is assumed that the dense inputs can either be broadcasted to or always have elements where the jagged inputs do.)

This suggests at least two possible approaches to the kernel implementation, both of which are included in this diff:

- In the **dense space indexing**, the kernel is launched on the `num_elements` in the dense output volume: the smallest rectangular volume that fits the jagged outputs (as well as all the dense and jagged inputs, by assumption of the front-end). The flat `dense_idx` for accessing the dense inputs is readily available from the CUDA indices (`blockIdx` / `threadIdx`). The flat `jagged_idx` for accessing the jagged inputs and outputs is computed from the `dense_idx` using the `offsets` of the jagged inputs / outputs (see the `KERNEL_COMPUTE_DENSE_IDX_THEN_JAGGED_IDX_TEMPLATE` for the details of the index computation). The `jagged_idx` computation requires accessing the `offsets` only once (at two consecutive positions), which leads to `O(1)` reading of the `offsets`. If it turns out that the jagged output does not have an element corresponding to a certain `dense_idx`, the thread returns prematurely.

- In the **jagged space indexing**, the kernel is launched on the `num_elements` in the jagged output shape: the `total_length` dimension followed by the dense dimensions in the jagged output (there can be more than one). The flat `jagged_idx` is readily available from the CUDA indices. The flat `dense_idx` is computed from the `jagged_idx` using the `offsets` of the jagged inputs / outputs (see the `KERNEL_COMPUTE_JAGGED_IDX_THEN_DENSE_IDX_TEMPLATE` for the details of the index computation). The `dense_idx` is computed using binary search within the `offsets` array, hence the `offsets` are read `O(log n)` times in each thread (with `n` being the length of the `offsets`).

The two approaches offer a trade-off:

- The dense space indexing runs in the dense output volume. Therefore, the number of threads may, in general, be much larger than the number of elements in the jagged output Tensors; and there may be many threads that return prematurely without doing any work. The index computation, on the other hand, is simple and lightweight.

- The jagged space indexing runs in the jagged output shape. As a result, the number of threads equals (up to the fixed gap of `FUSED_ELE_THREAD_SIZE`) to the number of elements in the jagged output Tensors; and (almost) each thread does meaningful work. But the binary search-based index computation may be heavy.

Reviewed By: ipiszy

Differential Revision: D43482363

fbshipit-source-id: 8194edda7145a839d53e0a581f24e1336dfdeb60