Use first dim as batch size if same in every input (#349)
Summary:
Pull Request resolved: https://github.com/facebookincubator/AITemplate/pull/349

If all input tensors have the same shape (in particular, the same first batch dimension), the current frequency calculation code can pick arbitrary dimension as `batch_size`.

This diff adds a special handling of the case when all inputs' first dimensions are the same. This guarantees that the first dimension is the most frequent one (or at least one of the most frequent ones). In such a case it is probably reasonable to declare the first input dimension as `batch_size` (because it typically is).

Reviewed By: tissue3

Differential Revision: D43716722

fbshipit-source-id: dae80c3abbad537fb60e100b60fb0f5649b2d64b