Implement expand op (#312)
Summary:
Pull Request resolved: https://github.com/facebookincubator/AITemplate/pull/312

### Implement expand operator CUDA backend

Adding CUDA backend implementation for expand: https://fburl.com/code/nb2mcsmg.

The operator semantic should be the same as the pytorch version https://fburl.com/fljywh6p.

#### Implementation

The previous expand operator was a no-op version, which only worked under very limited conditions, namely when it expanded just a single, already existing direction, and could be merged into a following elementwise op that supports tensor broadcasting.

This new version actually expands the tensor, supporting multiple expansion dimensions, dynamic shapes and adding dimensions, just like the pytorch version. There are three CUDA kernels implemented, one dealing with the general case, and two which are specialized in order to be faster in certain scenarios.

The pytorch version is in principle more effective, nevertheless, because in pytorch it just needs to create a new view on a source tensor with different read strides. As AIT has no general notion of strides for tensor dimensions, this is not a real option at the moment, unless we add that support to tensors and operators on them.

#### Further possible optimizations (not part of this PR )

 * When adding leading dimensions, this can be decomposed into writing an upper part of the tensor ( requiring strided reads or writes ) and then repeatedly copying that tensor ( which can be accomplished using effective sequential reads and writes and can utilize shared memory )
 * Further operator fusions should be possible
 * With all immutable dimensions, a more efficient implementation would be possible via loop unrolling and precalculation of strides etc.

Reviewed By: chenyang78

Differential Revision: D43419041

fbshipit-source-id: 84ec2c4716c3e21860d1d55807cf649ed543ba2e