Set the leftmost most frequent dim as batch size (#357)
Summary:
Pull Request resolved: https://github.com/facebookincubator/AITemplate/pull/357

If there is more than one most frequent dimension in the input shapes, the leftmost one: the one with the lowest position score (sum of position indices in the shapes) is picked as the batch size.

If there are multiple most frequent dimensions with the same position score, the choice is still arbitrary.

Reviewed By: wushirong

Differential Revision: D43755669

fbshipit-source-id: a8c10bbd2977e953ce44a22b0ee2df8e7c976963