Add jagged_to_dense op (#380)
Summary:
Pull Request resolved: https://github.com/facebookincubator/AITemplate/pull/380

In this diff, `jagged_to_dense` front-end and back-end op is added with the vectorized I/O. We reuse many of the utilities in `testing/jagged_utils.py` similar to `backend/common/elementwise_common.py`  in D43482363. A unit test and benchmark are included.

## Implementation Details
Since the output is dense, we adopt the calculations based on dense shape and apply padding when current element is outside of jagged shape from input.

Differential Revision: D43562375

fbshipit-source-id: ee80677ece18a2434d4751dfeff35ecbd47eb16b