Eliminate permute over singleton dims (#378)
Summary:
Pull Request resolved: https://github.com/facebookincubator/AITemplate/pull/378

Since when permute operation moves only a (or multiple) singleton dimension(s) memory layout does not change, it can be replaced with a reshape operation. New transform function for detecting and replacing operators in such cases is implemented. It's added among compiler optimization functions list. Unittests for this new functionality is added.

Reviewed By: chenyang78

Differential Revision: D43878517

fbshipit-source-id: 521017f1e8610bca07e40e6e63f5c39bb33df0d6