fix getitem issue
Summary:
Fix an issue in getitem. The operator.getitem stores the param in `args` so we need to copy it into `kwargs`. In this way, the downstream operator like slice are invoked in getitem and can use the `kwargs` w/o modifications.
If this diff is landed, we do not have to land D44005383 for reversion.

Differential Revision: D44007352

fbshipit-source-id: dc0781f637ae6f9132fa79772549e37a42095334