fix getitem issue (#411)
Summary:
Pull Request resolved: https://github.com/facebookincubator/AITemplate/pull/411

Fix an issue in getitem. The operator.getitem stores the param in `args` so we need to copy it into `kwargs`. In this way, the downstream operator like slice are invoked in getitem and can use the `kwargs` w/o modifications.
If this diff is landed, we do not have to land D44005383 for reversion.

Reviewed By: tissue3

Differential Revision: D44007352

fbshipit-source-id: a88c641e98b2a5f70161e4e866e2bfc52de69155