Add allow list to splitter
Summary:
Currently splitter only rejects nodes, but doesn't have allowed node list. Then there are some special ops hard to adapt to general rules.

Example:
torch.ops.fbgemm.jagged_dense_dense_elementwise_add_jagged_output should be supported after AIT adds it, but the second input of the op is offset, which is int64 type. But AIT in general only support float32 and float64 calculation.
There we just want to enable that op.

We added an allow_list to splitter so that if any op is in the allow_list, the splitter will split it into acc graph, rather than gpu graph.

Now, we can write
```
    class JaggedOperatorSupport(OperatorSupportBase):
        def is_node_supported(self, submodules, node: torch.fx.Node) -> bool:
            return node.op == "call_function" and node.target in [
                torch.ops.fbgemm.jagged_to_padded_dense,
                torch.ops.fbgemm.jagged_dense_dense_elementwise_add_jagged_output,
            ]

    op_lowering_disallow_list = []
    op_support = create_ait_operator_support(
        op_lowering_disallow_list=op_lowering_disallow_list,
        allow_op_supports=[JaggedOperatorSupport()],
    )

```
and we will get the same splitter but with one specific op get excluded!

Reviewed By: wushirong

Differential Revision: D44036284

fbshipit-source-id: a97c61c660420c89d439d0b8ebd1b1380f96cf4a