Remove redundant DeviceToDeviceCopies call
Summary: As `model->DeviceToDeviceCopies()` is called from the `ModelBase::Run`, in this diff we remove the redundant call from the end of the generated `Model::RunImpl()` function body in the `MODEL_TEMPLATE`. Redundant call has caused an overhead of unnecessary double copying of the same data in cases when the generated body of `DeviceToDeviceCopies` wasn't empty.

Differential Revision: D44074465

fbshipit-source-id: e9626844240e13c330ee9805d8a1d5065429f402