Remove redundant DeviceToDeviceCopies call (#429)
Summary:
Pull Request resolved: https://github.com/facebookincubator/AITemplate/pull/429

As `model->DeviceToDeviceCopies()` is called from the `ModelBase::Run`, in this diff we remove the redundant call from the end of the generated `Model::RunImpl()` function body in the `MODEL_TEMPLATE`. Redundant call has caused an overhead of unnecessary double copying of the same data in cases when the generated body of `DeviceToDeviceCopies` wasn't empty.

Reviewed By: chenyang78

Differential Revision: D44074465

fbshipit-source-id: 58b6a37062566e34a0787ef8438e5c2bd306c43b