Add dense_to_jagged op (#386)
Summary:
Pull Request resolved: https://github.com/facebookincubator/AITemplate/pull/386

This diff adds a new op, `dense_to_jagged`, to convert the input dense Tensor to the output jagged Tensor, given the offsets of the resulting jagged Tensor.

The op also takes the `total_length` dynamic (`IntVar`) dimension of the resulting jagged Tensor, as it is not inferable from the dense input Tensor's shape (only the `batch_dim` and jagged dims are). If the `total_length` is not initialized from any model input Tensor's shape, it is set within the `dense_to_jagged` back-end from the very last offset value in the very last offsets Tensor in the `offsets_list`.

Due to the current design aspect: the Tensors in the graph which are not outputs of any op and not model inputs are constant-folded, it doesn't seem possible to create the `source` of the resulting jagged Tensor "out of thin air" and pass it through the `make_jagged` op's back-end (validating the offsets content) *before* the `dense_to_jagged` op's back-end. As a workaround, we declare the `source` Tensor as the output of `dense_to_jagged`, then pass it (together with the `offsets_list`) through the `make_jagged`, and then return from the `dense_to_jagged.__call__` (see the details in the function body). Although this conforms to the above-mentioned aspect (i.e., `source` is not constant-folded, as it is, formally, the output of the `dense_to_jagged` op), we have to take a leap of faith in operating on not-yet-validated offsets Tensor content in the `dense_to_jagged` back-end. On the upside, the offsets are validated in the subsequent invocation of the `make_jagged`, before the resulting jagged Tensor can be processed further.

Differential Revision: D43944424

fbshipit-source-id: 193ae87583e80db2f7131feca8ee2b78624ff0b8