Make fuse_expand_elementwise jagged-aware (#426)
Summary:
Pull Request resolved: https://github.com/facebookincubator/AITemplate/pull/426

The `_fused_expand_elementwise` function within the `remove_no_ops` graph optimization pass eliminates `expand` ops preceding the `elementwise` ops. However, the elimination doesn't work with mixed jagged / dense `elementwise` inputs out of the box. Due to the fact that the `JaggedIntVar` in the jagged input shape is treated as a single dimension, against potentially more than one dimension in the dense input shape, the shapes don't match and the pass is not applied.

This diff expands the `JaggedIntVar` into its maximum dense shape. As a result, dense input shapes are comparable with the jagged shape "extended" like this. We rely on the mixed jagged / dense input compatibility invariant established by the `elementwise` op's front-end, and don't check it again in the pass.

Reviewed By: ipiszy

Differential Revision: D44063504

fbshipit-source-id: 8d9c04b230a838817797133ca14947c6ec29a84a