Set batch_dim in make_jagged if isolated (#436)
Summary:
Pull Request resolved: https://github.com/facebookincubator/AITemplate/pull/436

Previously, the `batch_dim` of the `JaggedIntVar` was set in the `make_jagged` back-end function when the value of `batch_dim` was zero. When the `batch_dim` is isolated (i.e., not set from any of the input shapes), this may be problematic when multiple batches with different batch size enter the same compiled model as inputs: the first batch will set the value of `batch_dim` by its size, but the very next batch with different size will trigger an error (because the value of `batch_dim` is no longer zero).

In this diff, the logic of setting the `batch_dim` value is changed to rely on the property of the symbolic graph (whether the dimension is isolated or not) instead of the rumtime (whether the value is zero). For a compiled model, this former property doesn't change in the runtime. Hence consecutive batches with different sizes can rewrite the value of of the `batch_dim` dimension in the runtime without triggering an error.

Reviewed By: chenyang78

Differential Revision: D44113704

fbshipit-source-id: bd4c6d350ff9c6955af9c8c5005ccf39ba651a46