Add logics in elementwise to handle the case that a jagged tensor is treated as a dense tensor in AIT
Summary:
ATT, it looks that it's possible that a jagged tensor is identified as a dense
tensor in fx2ait lowering if this jagged tensor is never used in any fbgemm
jagged operators. e.g. Adding two jagged tensors together do not require any
special fbgemm jagged operators. This diff handles this case for elementwise.

We need to think systematically on how to solve this issue for other operators
as well. i.e. When jagged_int_var.total_length() == dense_int_var, sometimes we
want them to be treated equally (e.g. when calculating the broadcasted dim),
but sometimes we want to always pick up the jagged_int_var (e.g. when inferring
output shapes).

Differential Revision: D44191261

fbshipit-source-id: e1e90793b3685c63c5129840a67d3bd460d5ec2e