Add dynamic JaggedDim bounds support (#452)
Summary:
Pull Request resolved: https://github.com/facebookincubator/AITemplate/pull/452

Previously, `JaggedDim`---a jagged dimension encoded in the first `total_length` / `sum_B(N_B)` dimension of a jagged Tensor---could only have static (`int`) `min_value` and `max_value`. This is problematic when the bounds (in particular, the upper bound: `max_value`) of a jagged dimension in a jagged Tensor shape, as well as the corresponding `max_seq_len` dimension in a dense Tensor shape, are dynamic (`IntVar`). The latter can happen, as the maximum sequence length can, in principle, be dynamic within a single compiled AIT model.

This diff promotes the `min_value` / `max_value` of `JaggedDim` to either `IntImm` for static cases (backward-compatible with previously supported `int` use cases) or `IntVar` for dynamic cases.

Reviewed By: frank-wei

Differential Revision: D44198794

fbshipit-source-id: 1650ba4832c62cd4c9ca923602117caacd5f16c6