Add full operator (#458)
Summary:
Pull Request resolved: https://github.com/facebookincubator/AITemplate/pull/458

This diff is adding a new operator: `full`. The op fills the output Tensor of a given `shape` and `dtype` with a (float scalar) `fill_value`. As the memory planning is done in the maximum shape of the op output, the kernel fills the maximum memory with the `fill_value` to allow constant-folding the output even if the shape contains `IntVar`s.

Differential Revision: D43909589

fbshipit-source-id: 92c18f6ed34499f57727ae145c966adb48b9f653