Add full operator (#458)
Summary:
Pull Request resolved: https://github.com/facebookincubator/AITemplate/pull/458

This diff is adding a new operator: `full`. The op fills the output Tensor of a given `shape` and `dtype` with a (float scalar) `fill_value`. As the memory planning is done in the maximum shape of the op output, the kernel fills the maximum memory with the `fill_value` to allow constant-folding the output even if the shape contains `IntVar`s.

Reviewed By: wushirong

Differential Revision: D43909589

fbshipit-source-id: 6c9c522efcbf183cacc5dd8fa4946c50fb65eb2b