Add flag to make_jagged to skip seq len check (#462)
Summary:
Pull Request resolved: https://github.com/facebookincubator/AITemplate/pull/462

Currently, `make_jagged` op's back-end always checks the offset differences to be within the corresponding `JaggedDim`'s bounds. Theoretically, there can be cases where a jagged Tensor can have sequences beyond the `JaggedDim.max_value()` (e.g., when the latter is defined by the corresponding dense Tensor's shape among the `elementwise` operands).

Here we add a flag to skip checking the sequence lengths in such cases (by default, the check is enabled).

Importantly, a guard is added to the `KERNEL_COMPUTE_JAGGED_IDX_THEN_DENSE_IDX_TEMPLATE` in `elementwise_common.py` to `return` prematurely from the CUDA thread handling the flat jagged index if the latter happens to be ouside the sequence bounds of the corresponding dimension of the dense volume. Previously, this was not needed, as the jagged sequences were always assumed to be covered by the dense volume.

Reviewed By: tissue3

Differential Revision: D44278023

fbshipit-source-id: 927ae99f81db20c3ee2a4165e4d6d4d42cdc12cf