Improve graph visualization (#457)
Summary:
Pull Request resolved: https://github.com/facebookincubator/AITemplate/pull/457

Creates an additional graph visualization in case of profiling. Becomes available once `LOGLEVEL=DEBUG` is used, and it is recommended to enable  `AIT_PLOT_SHORTEN_TENSOR_NAMES=1` as well.

Highlights ops with the duration higher than 70% percentile.

Every op gets marked with an execution duration. Size is proportional to the duration for the above 70% percentile.

Every tensor gets two timestamps in the form of 'A ms / B ms'. A is the timestamp when the tensor becomes computed using a single stream execution. B is the timestamp when the tensor becomes computed using unlimited number of independent streams, thus serving as a theoretically minimal timestamp at which the tensor may be computed.

The ratio of `A/B - 1` for the output tensor is how much speed improvement one may gain in case of unlimited amount of GPUs available and zero copy cost.

Reviewed By: ipiszy, chenyang78

Differential Revision: D44183018

fbshipit-source-id: 8568d06dd090086d743fc6c0c4eb68ddb3d2d392