Resurrect gemm_ / bmm_softmax
Summary:
The `gemm_` / `bmm_softmax` ops are broken for ~10 months, with the corresponding tests disabled, due to an inconsistency with the CUTLASS API update in 2.10. As a result, the major changes in the GEMM back-end did not cover these ops.

In this diff, the following `softmax` ops are resurrected:

- `gemm_rcr_softmax`
- `gemm_rcr_bias_softmax`
- `bmm_rcr_softmax`

More specifically, the CUTLASS API inconsistency is resolved, the ops' GEMM back-end code is modernized to be compatible with the outer GEMM ecosystem, the unit tests are made functional again. Importantly, the diff introduces minimal changes necessary to make the above `softmax` ops functional again. Further improvements (some listed below) are *not introduced* here.

What is *done* in this diff (based on terrychenism's [unpublished branch](https://github.com/fairinternal/AITemplate/compare/main...terrychenism:AITemplate:softmax_update) from 8 months ago):

- Update the internal `GemmSoftmaxUnversal` CUTLASS-based operator to be consistent with the 3.0 APIs.
- Fix the bug in `GemmSoftmaxUnversal` operator (also present in the upstream `GemmSoftmax`).
    - To allow SMEM > `(48 << 10)` bytes required by some of the generated op instances.
- Use op workspace instead of *all* temporary inputs.
- Generalize `dtype` instead of the hard-coded `half`.
- Combine profilers into a single file, as required by the outer `gemm_universal` code.
- Remove `split_k` support from the codegen, as the CUTLASS op uses `kBatched` mode hard-coded.

What is *not done* in this diff (hence, should probably be done in the future):

- Performance tuning (e.g., based on tweaking the `ApplyShape` as suggested [here](https://github.com/NVIDIA/cutlass/blob/master/examples/35_gemm_softmax/gemm_softmax.cu#L221-L224)).
- Use upstream CUTLASS [`GemmSoftmax`](https://github.com/NVIDIA/cutlass/blob/master/examples/35_gemm_softmax/gemm_with_softmax.h#L317) operator instead of the internal `GemmSoftmaxUniversal`.
    - Due to the SMEM bug. Will submit a PR to `nvidia/cutlass` soon, can switch when merged.
- Support arbitrary number of `a` dimensions instead of M in `gemm_` ops.
- Support input TensorAccessors (output TAs seem to be supported).
- Other things that are possibly missed.

Differential Revision: D44406815

fbshipit-source-id: c9df727e29041e9d6aaa06274cba88ee3e62d468