Introduce safe_compile_model with locking
Summary:
This diff extends D44374161 in the following ways:

one criticism with respect to compile_model locking it's build dir is, that this is kind of unexpected behavior and not part of the function's core functionality.

So I factored out the locking into an outer function called **safe_compile_model** which can act as a drop-in replacement for compile_model, but locks the build_directory and then delegates to compile_model.

Additionally, all references to compile_model in unit tests were replaced by references to safe_compile_model.

Differential Revision: D44421367

fbshipit-source-id: 34070a0bbf8c92d597f86fab931a4033e612da0c