AIT build cache (#456)
Summary:
Pull Request resolved: https://github.com/facebookincubator/AITemplate/pull/456

The compilation stage of AIT's generated source files takes a considerable amount of time. It would be nice if we could speed up the build using a build cache, for example using ccache. This would be especially beneficial for speeding up
Unit tests, possibly with a huge cost benefit to Meta,
considering how much time is spent in AIT unit tests on GPU build servers ( see https://www.internalfb.com/sevmanager/view/322359 ) - also, we as developers might wait unneccessarily long for unit tests to complete, so developer productivity could be improved.

Unfortunately, ccache does not work with AIT build scripts for nvcc as it destroys the options that need to be passed down to the host compiler ( I tried that).

But luckily, AIT's codegen has a nice property that we can exploit: The generated sources are rather compact, and all in a single directory. So instead of caching build results for a single file, we could attempt to hash *all* of the sources of a build directory, and if nothing changed, to retrieve the build results for that directory instead of invoking make.

This would likely speed up CI builds considerably.

This diff has a small POC to test this hypothesis.

I was able to, for example, bring down the execution time of this small set of unit tests down from approx. 60 to 15 seconds  after having warmed the build cache:

  buck2 run @//mode/dev-nosan AITemplate:test_gemm --local-only -- -r 3d_2d_rcr

or this set, from approx. 530 seconds, down to approx. 100 seconds.

  buck2 run @//mode/dev-nosan AITemplate:test_expand --local-only

It works rather simple: Define an environment variable
like this:

  export AIT_BUILD_CACHE_DIR=/tmp/ait_build_cache

and that's it. If it is not defined, it is disabled.

### Note:

This is a quick PoC to test the idea and prove we can speed up our CI and dev workflow considerably - to actually productionalize it on our CI, a few points would have to be taken care of:

### TO DO

for production readiness in CI ( in followup diff(s)?)

 * ~~Old cache entries need to be cleaned up~~ (Done!)
 * ~~Make cache updates atomic ( write to temporary directory, then rename to final directory name = hash ) to prevent race conditions in distributed / parallel builds~~ (Done!)
 * ~~Ensure that no environment variables play a role in the build (beyond being included in the Makefile verbatim) - if they do, include in hash~~ ( Checked and tested manually)
 * ~~Filter html, json and possible other debug / trace info, so that these do not contribute to hash calculation~~ (Done!)
 * Run entire test suite twice, and see if there are tests where the cache does not hit on second run ( if yes, investigate why -> Likely the code generation was not generating exactly the same sources twice )
 * ~~(Store the cached data in a distributed manner on the build servers ( network fs, memcached? ccache seems to use memcached there ) and that it is properly expired.~~) (Separate task created)
 * ~~Integrate with CI build system~~ (Separate task created)
 * ~~Add unit tests, which prevent regressions ( e.g. allow to check that future changes in code generation do not break the caching mechanism, for example by not generating deterministically the same code on repeated invocation with same inputs).~~(Done)
 * ~~Ensure that differences in compiler versions are reflected in the hash, so either ensure that used compiler versions are part of the path they are being referred to in the Makefile, or hash these versions explicitly ( they could just be written to a file in the build directory that's included in the hash by executing something like "nvcc --version >>compiler_versions.hashinfo" etc)~~ (Done)

Reviewed By: chenyang78

Differential Revision: D44229622

fbshipit-source-id: 7a767343909e08d9cb806f50698b510bd917e3f2