Accept multiple sources in make_jagged (#519)
Summary:
Pull Request resolved: https://github.com/facebookincubator/AITemplate/pull/519

`make_jagged` currently accepts only a single `source` and `offsets_list`, returning a jagged Tensor made from those. This may be inefficient for the cases where there are multiple jagged Tensors in the graph with the same `total_length` dimension / same offsets (e.g., in the group b2b bmm). In this case, `make_jagged` would have to be invoked as many times, doing the same offset validation on the GPU in every run. The latter is a waste of time.

In this diff, `make_jagged` is extended to accept multiple source Tensors (with the same first `total_length` dimension) and convert each to a jagged Tensor.

The extension is necessary for the upcoming graph transformation pass eliminating multiple `make_jagged` ops within the graph by pulling them upstream / applying `make_jagged` to the sets of source input Tensors with the same `total_length` first dimension. The pass will be added in a follow-up diff.

Reviewed By: chenyang78

Differential Revision: D44508562

fbshipit-source-id: 9424ae3428b2b05a0bea04af7c7c42d2e5de0a92