Skip acc normalization of repeat_interleave if input dims aren't integral (#528)
Summary:
Pull Request resolved: https://github.com/facebookincubator/AITemplate/pull/528

There's cases where we have torch.repeat_interleave(x, repeats=y), where the y is a tensor. In these cases, don't do the mapping since we can't run tile on those dims.

To do this, skip creating a new replacement node in the mapper, and add an option to skip the node mapping in acc_normalizer called "skip_normalization_if_none". If true, then following normalization the returns none, it'll set the args/kwargs back so that the graph recompilation can succeed.

Differential Revision: D44568748

fbshipit-source-id: 942d49643435fd2c12aea8fa3620132934733cd2