Support simple variable_seq_length in FMHA_style_b2b_bmm (#486)
Summary:
Pull Request resolved: https://github.com/facebookincubator/AITemplate/pull/486

ATT, the block size / thread  size is still based on max_seq_length, but skip elements which are not within actual_seq_length.
It's not as optimal as the solution with group scheduling algorithm (especially when seq_lengths differ a lot for different examples in a batch), but it could help avoid jagged / padded conversion and should work fine when seq_lengths do not differ a lot.

Reviewed By: aakhundov

Differential Revision: D44364087

fbshipit-source-id: 2ca7f96e2a02f3edb660b61af7cc258aba99139b