add priority toposort (#408)
Summary:
Pull Request resolved: https://github.com/facebookincubator/AITemplate/pull/408

Following the suggestion in T147269828,
Use Kahn's algorithm with a priority queue, the priority definition of the priority queue would be the order of the sorted graph.

Following this note: https://fb.workplace.com/notes/1322142898565804 , added a "SizePriTensorHelper" , which larger size tensor would be in the top rank of the sorted graph.

In the future if we'd like to test different types of priority, simply add a new class inherited from "PriTensorHelper", overwrite the function "get_priority"

Reviewed By: tenpercent

Differential Revision: D43994537

fbshipit-source-id: 95012bc231e1212d36ce3481349212e0faa6a942