Carry over check_sequence_lengths in dedup_make_jagged_ops pass
Summary: The `check_sequence_lengths` attribute of the `make_jagged` ops is not carried over in the `dedup_make_jagged_ops` from the old ops to the new one. This is a bug that causes problems in the setting where `check_sequence_lengths` is set to `False` in the existing ops, as the default value is `True`. The diff fixes the bug by carrying over the attribute in the pass.

Differential Revision: D44808132

fbshipit-source-id: 4dec19483dc67690e8a1de2185b79b2d72bf41f2