Carry over check_sequence_lengths in dedup_make_jagged_ops pass (#549)
Summary:
Pull Request resolved: https://github.com/facebookincubator/AITemplate/pull/549

The `check_sequence_lengths` attribute of the `make_jagged` ops is not carried over in the `dedup_make_jagged_ops` from the old ops to the new one. This is a bug that causes problems in the setting where `check_sequence_lengths` is set to `False` in the existing ops, as the default value is `True`. The diff fixes the bug by carrying over the attribute in the pass.

Reviewed By: amateurcoffee, tissue3

Differential Revision: D44808132

fbshipit-source-id: 5bdd8d8d764cafbf06e0bacfaffe973db2aecf25